requirements.txt
Django==4.2.2
django-admin==2.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
mysqlclient==2.1.1
Pillow==9.5.0
PyJWT==2.7.0
python-dotenv==1.0.0
rest-framework-simplejwt==0.0.2

